Pagination Page navigator at the bottom Redirection from update status to view status
