+++ src/main/webapp/WEB-INF/tiles.xml
@@ -18,4 +18,7 @@
     <definition name="app.addStatus" extends="app.default">
         <put-attribute name="content" value="/WEB-INF/tiles/addstatus.jsp" />
     </definition>
+    <definition name="app.viewStatus" extends="app.default">
+        <put-attribute name="content" value="/WEB-INF/tiles/viewstatus.jsp" />
+    </definition>
 </tiles-definitions>
