+++ src/main/webapp/css/main.css
@@ -62,3 +62,28 @@ textarea {
 .errors {
     color: red;
 }
+
+.status-list-id{
+    width: 5rem;
+}
+
+.status-list-status{
+    width: 15rem;
+}
+
+.status-list-date {
+    width: 7rem;
+}
+
+.pagination{
+    margin-top: 20px;
+    margin-bottom: 30px;
+    margin-left: auto;
+    margin-right: auto;
+    width: 10%;
+}
+
+.pagination-num {
+    margin-right: auto;
+    margin-left: auto;
+}
